requirements.txt
Django==1.9.7
djangorestframework==3.3.3
git+https://github.com/2checkout/2checkout-python.git@3.x
